Ajuste para que limpasse a lista de tarefas após clicar no botão Resetar tarefas

+++ static/resetTasks.js
@@ -1,18 +1,25 @@
 document.getElementById('resetTasks').addEventListener('click', function () {
-    fetch('/resetar_tarefas', {
-        method: 'DELETE',
-    })
-        .then(response => {
-            if (!response.ok) {
-                throw new Error('Network response was not ok');
-            }
-            return response.json();
+    if (confirm('Tem certeza que deseja resetar todas as tarefas?')) {
+        fetch('/resetar_tarefas', {
+            method: 'DELETE'
         })
-        .then(data => {
-            alert(data.message);  // Mostra a mensagem de sucesso
-            document.getElementById('taskList').innerHTML = '';  // Limpa a lista de tarefas na interface
-        })
-        .catch(error => {
-            console.error('Erro ao resetar as tarefas:', error);
-        });
+            .then(response => {
+                if (response.ok) {
+                    alert('Tarefas resetadas com sucesso');
+
+                    // Limpa todas as listas de tarefas no front-end imediatamente
+                    document.getElementById('todoList').innerHTML = '';
+                    document.getElementById('doingList').innerHTML = '';
+                    document.getElementById('doneList').innerHTML = '';
+
+                    // Exibe mensagem de sucesso
+
+                } else {
+                    alert('Erro ao resetar tarefas.');
+                }
+            })
+            .catch(error => {
+                console.error('Erro ao resetar tarefas', error);
+            });
+    }
 });
